chore(integrations/intercom): extract channels in dedicated file (#14434)
